Add caching on get_test_runs_by_build_name
This commit enables simple caching on the results from the
get_test_runs_by_build_name. This API call involves one of the slowest
subunit2sql DB calls we make and is a good candidate for caching.
Right now, we just cache on name and the request datestamps which will
result in a lot of potential cache misses as people request varied
dates and/or build_names. So there is probably room for additional
optimizations in the future.

Change-Id: I7582188f4f027bc86375056e5cf60cca3a760ce8