Add total count table to per test page
This commit adds a table to the bottom of the per test page that
details the total counts and percentages by status for a particular
test over the selected time window. While the area plot is good to
visualize this overtime, it's not good for figuring out aggregate
data, which is where the table is useful.

Change-Id: I58b0d834fb8e7d74ce1aa7a8020f54a238995ba4