Change pass-fail graph to multi bar chart
This commit changes the pass-fail graph on test page to multi bar chart
graph. I think this is better than the historical bar chart. But the
header of the tooltip is not correct when the flag of
chart.useInteractiveGudeline(true) due to this issue[1]. So I just set
it false as a workaround. We may need to set it true if we want to see
in one tooltip for each data.

[1] https://github.com/novus/nvd3/pull/1722

Closes-Bug: #1611105
Change-Id: I9e418b77eb03956be16de45df701ae94ea762cde