Cache individual elasticsearch calls
When we introduced elastic-recheck querying to openstack-health we
started caching the entire api call whenever es was being queried on
the backend. At the time this made sense because there was a single
api call using elastic search so this was easier. But now that we have
multiple apis using elastic search data we should individually cache
the calls. There are 2 reasons for doing this:

1. We can share the cache of queries across api calls. We only use ES
to run e-r queries on failed runs. Even if looking at fialures from a
different view there will be a lot of overlap in what we're searching
for.

2. With the addition of failure lists that depend on a date range our
cache hit rate goes way down for the api call. That's because our
caching strategy is a bit naive still and doesn't understand ranges so
we always have a cache miss if one of the bounds on the range misses.
To mitigate this by caching ES queries the slow part of the query will
always be cached except for anything new.

Change-Id: Ibf1589e43598dba92e6b6e36f481c38ecc5116ee