Fix bad use of jsonify() in _get_data
When no test runs are found in the subunit2sql database we attempt to
return an empty result set in get_test_runs_for_test. However, in the
cached function _get_data() the empty response was wrapped in a
call to jsonify() causing the function to return a Response rather
than a serializable dict.

This patch removes the bad jsonify() call to avoid any serialization
errors for pages with no data.

Change-Id: I49d47bb998593f61fc4b1431d366c4cdd9bb3d2c