Add canvas scatter chart
This commit adds a canvas-based scatter chart using the new canvas
directives, replacing the nvd3 scatter/line multi-chart on the test
page.

Note that this chart also replaces the scatter chart as well as the
multi-chart as all canvas charts are implicitly multi-charts: we can
draw as many charts on top of one another as we'd like.

Change-Id: I3c560f1e04631c1e72aecac4bf0af656a04a3d6f