Fix elasticsearch querying with recent e-r changes
A recent elastic-recheck commit, change
Ic6f115a6882494bf4c087ded4d7cafa557765c28 made backwards
incompatible changes to the api for the classfier object to use a
common config object. This was a good cleanup but failed to account
for things reusing elastic-recheck as a lib interface outside of the
repo (like openstack-health). This change refactors the
openstack-health usage to take into account these recent changes and
unbreak things.

Change-Id: I3afe7f0a4a4475084be54f8df69083105fb4a4fe