Add version string to status string
Sometimes debugging puppet deployment issues is difficult without
knowing exactly which version of the code is actually running. To help
debug automated deployment issues being able to query the exact version
running is useful. This commit does this by adding the version string to
the /status response so we can easily tell which version is running.

Change-Id: I4a26cf097cd1efc9962911e7b5824eb6b8b2f5d9