Update angularjs libraries version
This commit updates angularjs libraries version. This can be included in
the other npm packages update patch, however, I took a deliberate way
this time to avoid accident as possible.
This version is the latest against AngularJS[1] but not Angular[2]
because Angular is totally different from AngularJS. It's not so easy to
upgrade to it.

[1] https://www.angularjs.org/
[2] https://angular.io/

Change-Id: I0941d9a537681659b17150fc532668d6bc12fc05