Trust resource URLs in health-api for various environments
This commit makes resource the URLs in health-api.js for various
environments such as in local development environments, downstream
environments. The URLs should be trusted because the apiRoot value comes
from the config.
If you'd like to know more details of $scedelegateprovider, you can
see the documents at angularjs site[1].

[1] https://docs.angularjs.org/api/ng/provider/$sceDelegateProvider

Change-Id: Iec67c1a5236c9c8febfc83188db71754f895bff6