Fix config file path treatment when using uwsgi
This commit fixes the treatment of the config file path when we use the
Flask app directly. Originally, when we use the Flask app such as using
with uwsgi which is mentioned in the README as one of production
settings, the parameter '--pyargv config_file' doesn't work because we
read '/etc/openstack-health.conf' file directly. It should work,
otherwise, users can't specify the config file.

Change-Id: I1beb927200880d42e40e49db3aab47453f83169f