Fix npm-test: write after end error
The npm-test gate currently fails while bundling due to a
write after end error.
Here are some more details on the error:
  *snippet of the error[1]
  *link to sample run[2]

I tried a few solutions online and found this[3] to work with
how we bundle openstack-health. The release for 1.4.2 has been
out for some time so I was curious why it's just now starting to
fail but at least this gets the gate passing again.

[1]http://paste.openstack.org/show/688811/
[2]http://logs.openstack.org/66/547866/7/check/nodejs-npm-run-test/f21e7b5/
[3]https://github.com/substack/bulkify/issues/18

Change-Id: Id288e68a7b38382d99a1dfb886544f99c0714310