Make the docker files more efficient
Copy dependencies from the context and install them before
copying the app code and installing it.  Do not reinstall
all dependencies when app code changes.

Define dedicated .dockerignore files to allow the FE to ignore
changes in the API and viceversa. Define extra params / steps
in the source to image task to deploy a dedicated dockerignore.