Fix gate: define nodejs10 jobs
Openstack-health uses horizon nodejs template [1] to run
nodejs10- jobs Which is wrong because openstack-health is
not a horizon plugins and above template is only for horizon
plugins to run node-js jobs.
This patch remove horizon nodejs template from .zuul.yaml
and define a new job to run Nodejs10 jobs on gate.

Also drop support of py35.

[1] https://opendev.org/openstack/openstack-health/src/branch/master/.zuul.yaml#L3

Change-Id: Ia222ce7f52ee1cfd575ed5204088f0eee2fe903e