Retire openstack-health
openstack-health is QA repo and service to know the
job/test success/failure rate. But for a long (more than
a year), this code is broken. We do not have any maintainer
in QA to fix and maintain it. As it is broken, QA and infra team
agree to stop the service

- http://status.openstack.org/openstack-health/#/

In QA zed PTG, we decided to retire the repo also

- https://etherpad.opendev.org/p/qa-zed-ptg

Needed-By: https://review.opendev.org/c/openstack/governance/+/836706
Change-Id: Ie15aa8e469b0bb3aff47dcad422c0676fea640d2